fix: make sure HelmetData uses the same context data under the hood
RSC cannot serialize methods and functions between the server and client,
so we need to support recreating `HelmetData` with the serialized data
from the server. This works because `HelmetData` itself is just a wrapper
over a context object and an instances array. Those can be the same refs
in memory for multiple HelmetData instances.